Added more lifecycle wrappers...
The `componentWillMount` and `componentWillUpdate` are helpful if you
need to setup/cleanup event handlers outside of React. An example would
be a component that listens for history events to do some routing. Or a
component that listens for key events on the document or window object
to hide or show a modal dialog.